requirements.txt
Django==3.0.3
djangorestframework==3.11.0
django-cors-headers==3.10.1
pymongo==3.12.1
asgiref==3.2.3
djongo==1.3.6
pytz==2019.3
sqlparse==0.2.4
gunicorn==20.0.4
pymongo==3.12.1
dnspython==2.1.0
